requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
# backports.zoneinfo==0.2.1
billiard==4.2.0
celery==5.3.6
cfgv==3.4.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
distlib==0.3.7
Django==4.2.8
django-celery-beat==2.5.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
factory-boy==3.3.0
Faker==21.0.0
filelock==3.13.1
identify==2.5.33
importlib-resources==6.1.1
inflection==0.5.1
isort==5.13.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kombu==5.3.4
nodeenv==1.8.0
pkgutil_resolve_name==1.3.10
platformdirs==4.1.0
pre-commit==3.5.0
prompt-toolkit==3.0.43
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.32.0
rpds-py==0.13.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.12
zipp==3.17.0
